Se realiza cambios en TPROFORMAFACADE, se elimina archivos con contenido duplicado.
