--- sistemaProformasOnlineIGMv1/sistemaProformasOnlineIGMv1-web/src/main/webapp/tproforma/subeArchivo_1.xhtml
@@ -1,64 +1,0 @@
-<?xml version='1.0' encoding='UTF-8' ?>
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
-      xmlns:h="http://xmlns.jcp.org/jsf/html"
-      xmlns:p="http://primefaces.org/ui">
-
-    <h:head>
-        <title><ui:insert name="title">Default Title</ui:insert></title>
-        <h:outputStylesheet library="css" name="jsfcrud.css"/>
-        <h:outputScript library="js" name="jsfcrud.js"/>
-    </h:head>
-
-    <h:body>
-
-<!--        <p:growl id="growl" life="3000" />-->
-
-        <p:layout fullPage="true">
-            <p:layoutUnit position="north" size="250" header="Envío del comprobante de pago (Cliente)">
-                <h:form id="TproformaSendForm" enctype="multipart/form-data">
-                    <p:growl id="messages" showDetail="true" />   
-                    <p:outputLabel value="CI/RUC: " style="font-weight: bold" />
-                    <p:outputLabel value="#{login.aliasBase}" />
-                    <br />
-                    <p:outputLabel value="Nombre cliente: " style="font-weight: bold"/>
-                    <p:outputLabel value="#{login.cliente.nombres.concat(login.cliente.apellidos)}" />
-                    <br />
-                    <p:outputLabel value="No. proforma: " style="font-weight: bold"/>
-                    <p:outputLabel value="999" />
-                    <br />
-                    <br />
-                    <p:outputLabel value="Adjuntar archivo: "  
-                                   style="font-weight: bold"/>
-                    <p:fileUpload value="#{fileUploadView.file}"
-                                  mode="advanced" 
-                                  dragDropSupport="false"
-                                  fileUploadListener="#{fileUploadView.upload}"
-                                  uploadLabel="Subir imágen" 
-                                  sizeLimit="51200" 
-                                  cancelLabel="Cancelar" 
-                                  label="Elegir"
-                                  update="messages"/>
-                    <br />
-                    <br />
-                    <p:outputLabel value="#{TproformaController.p.tiempoEntrega}"/>
-                    <br />
-                    <br />
-                    <p:outputLabel value="El formato del archivo de envío debe ser en formato PDF de un máximo de 50 KB"  />
-                </h:form>
-            </p:layoutUnit>
-
-            <p:layoutUnit position="south" size="60">
-                <ui:insert name="footer"/>
-            </p:layoutUnit>
-
-            <p:layoutUnit position="center">
-                <ui:insert name="body"/>
-            </p:layoutUnit>
-
-        </p:layout>
-
-    </h:body>
-
-</html>
